requirements.txt
# general modules
django >3.2,<3.3
python-decouple
pillow
djangorestframework


drf-yasg[validation]